// Models/CanadaNationalStats.ts
interface CanadaNationalStats {
    CasesNew: number;
    CasesTotal: number;
    Recovered: number;
    DailyRecovered: number;
    TotalDeaths: number;
    DailyDeaths: number;
    TotalTested: number;
    DailyTested: number;
    TotalActive: number;
    DailyActive: number;
    TotalHospitalized: number;
    DailyHospitalized: number;
    TotalICU: number;
    DailyICU: number;
    TotalVaccinated: number;
    DailyVaccinated: number;
}

export default CanadaNationalStats;